+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Changelog
 - Fix issues with filesizes not showing correctly when downloading mods for a pack
 - Fix issues with username changes causing users to not be able to login with Mojang anymore
 - Add in --force-offline-mode command line argument to, well, force offline mode. Should work fine
+- Fix semi public packs being added/removed not refreshing the packs list